Add websocket support
