+++ README.md
@@ -1,8 +1,8 @@
 # Aioautomower
 
-## Asynchronous library to communicate with the Automower Connect API
+Asynchronous library to communicate with the Automower Connect API
 
-## This library is under development.
+## REST API Examples
 
 ```python
 from aioautomower import GetAccessToken, GetMowerData, Return
@@ -15,7 +15,7 @@ username = "username" ## Your username
 password = "password" ## Your password
 
 
-class Example_Token:
+class ExampleToken:
     """Returns the access token as dict."""
     def __init__(self, api_key, username, password):
         self.api_key = api_key
@@ -34,7 +34,7 @@ class Example_Token:
             raise KeyError
         return access_token_raw
 
-class Mower_Data:
+class MowerData:
     """Returns the data of all mowers as dict."""
     def __init__(self, api_key, access_token, provider, token_type):
         self.api_key = api_key
@@ -80,7 +80,7 @@ class SendingCommand:
         return send
 
 
-example = Example_Token(api_key, username, password)
+example = ExampleToken(api_key, username, password)
 token_output = asyncio.run(example.token())
 print(token_output)
 
@@ -88,7 +88,7 @@ access_token = token_output["access_token"]
 provider = token_output["provider"]
 token_type = token_output["token_type"]
 
-example2 = Mower_Data(api_key, access_token, provider, token_type)
+example2 = MowerData(api_key, access_token, provider, token_type)
 mower_output = asyncio.run(example2.mowers())
 print(mower_output)
 
@@ -96,4 +96,34 @@ mower_id = mower_output["data"][0]["id"] ## '0' is your first mower
 print (mower_id)
 payload = '{"data": {"type": "ResumeSchedule"}}'  ## For more commands see: https://developer.husqvarnagroup.cloud/apis/Automower+Connect+API#/swagger
 SendingCommand(api_key, access_token, provider, token_type, mower_id, payload)
+```
+
+## AutomowerSession examples
+
+An AutomowerSession keeps track of the access token, refreshing it whenever
+needed and monitors a websocket for updates, whose data is sent to callbacks
+provided by the user.
+
+```python
+import asyncio
+import logging
+
+import aioautomower
+
+USERNAME = "user@name.com"
+PASSWORD = "mystringpassword"
+API_KEY = "12312312-0126-6222-2662-3e6c49f0012c"
+
+
+async def main():
+    sess = aioautomower.AutomowerSession(API_KEY, USERNAME, PASSWORD)
+    await sess.connect(status_cb=lambda websocket_data: print(websocket_data),
+                       positions_cb=lambda websocket_data: print(websocket_data),
+                       settings_cb=lambda websocket_data: print(websocket_data))
+    await asyncio.sleep(5)
+    status = await sess.get_status()
+    print(status)
+    await asyncio.sleep(10000)
+
+asyncio.run(main())
 ```
